split out some color_correction functions for easier imports and clarity
