+++ plantcv/plantcv/transform/__init__.py
@@ -1,19 +1,19 @@
-from plantcv.plantcv.transform.color_correction import get_color_matrix
-from plantcv.plantcv.transform.color_correction import get_matrix_m
 from plantcv.plantcv.transform.color_correction import calc_transformation_matrix
 from plantcv.plantcv.transform.color_correction import apply_transformation_matrix
 from plantcv.plantcv.transform.color_correction import save_matrix
 from plantcv.plantcv.transform.color_correction import load_matrix
 from plantcv.plantcv.transform.color_correction import correct_color
 from plantcv.plantcv.transform.color_correction import create_color_card_mask
-from plantcv.plantcv.transform.color_correction import std_color_matrix
-from plantcv.plantcv.transform.color_correction import astro_color_matrix
 from plantcv.plantcv.transform.color_correction import affine_color_correction
 from plantcv.plantcv.transform.detect_color_card import deltaE
 from plantcv.plantcv.transform.detect_color_card import detect_color_card
 from plantcv.plantcv.transform.detect_color_card import mask_color_card
+from plantcv.plantcv.transform.get_color_matrix import get_color_matrix
+from plantcv.plantcv.transform.get_color_matrix import get_matrix_m
 from plantcv.plantcv.transform.rescale import rescale
 from plantcv.plantcv.transform.rotate import rotate
+from plantcv.plantcv.transform.standard_matrices import std_color_matrix
+from plantcv.plantcv.transform.standard_matrices import astro_color_matrix
 from plantcv.plantcv.transform.nonuniform_illumination import nonuniform_illumination
 from plantcv.plantcv.transform.resize import resize, resize_factor
 from plantcv.plantcv.transform.warp import warp, warp_align
